chore: changing kafka configs

+++ auth-user-service/src/main/resources/db/migration/V1__User_entity.sql
@@ -2,7 +2,7 @@ CREATE SEQUENCE IF NOT EXISTS hibernate_sequence START WITH 1 INCREMENT BY 1;
 
 CREATE TABLE users
 (
-    id       BIGINT NOT NULL,
+    id       SERIAL NOT NULL,
     email    VARCHAR(255),
     name     VARCHAR(255),
     phone    VARCHAR(255),
